Enhanced CollaborativeFilter with prototype scope
The CollaborativeFilter class has been updated to use the prototype scope from Spring. This change allows for a new instance of the filter to be created each time it's requested. Additionally, a constructor was added that prints a message when an instance is created, aiding in debugging and tracking object creation.