Added content-based movie recommendation feature
Implemented a new feature for providing movie recommendations based on content. This includes the creation of a ContentBasedFilter class, which implements a Filter interface and provides recommendations based on similar movies. Also added a Movie class with properties like id, name, genre, and producer. The MovieRecommenderSystemApplication now retrieves singleton and prototype beans from the application context to demonstrate the functionality of this new feature.